refactor(iwlwifi): split driver lifecycle modules
